LayoutTests/imported/w3c/web-platform-tests/html/browsers/browsing-the-web/navigating-across-documents/javascript-url-return-value-handling-dynamic-expected.txt


FAIL 0041 set in src="" assert_equals: expected "UTF-8" but got "windows-1252"
FAIL 0041 set in href="" targeting a frame and clicked assert_equals: expected "A" but got ""
FAIL 0080 00FF set in src="" assert_equals: expected "UTF-8" but got "windows-1252"
FAIL 0080 00FF set in href="" targeting a frame and clicked assert_equals: expected "ÿ" but got ""
FAIL 0080 00FF 0100 set in src="" assert_equals: expected "UTF-8" but got "windows-1252"
FAIL 0080 00FF 0100 set in href="" targeting a frame and clicked assert_equals: expected "ÿĀ" but got ""
FAIL D83D DE0D set in src="" assert_equals: expected "UTF-8" but got "windows-1252"
FAIL D83D DE0D set in href="" targeting a frame and clicked assert_equals: expected "😍" but got ""
FAIL DE0D 0041 set in src="" assert_equals: expected "\ufffdA" but got "U+de0dA"
FAIL DE0D 0041 set in href="" targeting a frame and clicked assert_equals: expected "\ufffdA" but got ""
